refactor: move UIState files from category to explore package
